[tf.data] graduate experimental unique api to tf.data.Dataset
